+++ Rave/Rave/Engine/Graphics/Include/BufferView.h
@@ -18,7 +18,7 @@ namespace rave
 			srvDesc.Format = format;
 			srvDesc.ViewDimension = D3D11_SRV_DIMENSION_BUFFER;
 			srvDesc.Buffer.FirstElement = 0;
-			srvDesc.Buffer.NumElements = data.size();
+			srvDesc.Buffer.NumElements = (uint32_t)data.size();
 
 			ResourceView::Load(gfx, buffer, srvDesc);
