+++ Rave/Rave/Engine/Graphics/Include/IndexBuffer.h
@@ -7,9 +7,9 @@ namespace rave
 	class IndexBuffer : public Buffer
 	{
 	public:
-		IndexBuffer& Load(Graphics& gfx, const std::vector<size_t>& vertices, const bool writeAccess = false);
+		IndexBuffer& Load(Graphics& gfx, const std::vector<uint32_t>& vertices, const bool writeAccess = false);
 
-		void Write(Graphics& gfx, const std::vector<size_t>& vertices);
+		void Write(Graphics& gfx, const std::vector<uint32_t>& vertices);
 
 		void Bind(Graphics& gfx) const noexcept;
 	};
